requirements.txt
pandas
streamlit
openai
matplotlib
python-dotenv
pandasai==0.6.10